# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Template API",
    "description": "Users, organizations, and todos.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "validation",
              "authentication",
              "forbidden",
              "not_found",
              "rate_limit",
              "internal"
            ],
            "description": "The error type category",
            "examples": [
              "validation",
              "authentication"
            ]
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code identifier",
            "examples": [
              "invalid_parameter",
              "missing_required_field",
              "unauthorized"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "examples": [
              "The request was invalid",
              "Authentication required"
            ]
          },
          "param": {
            "description": "The parameter that caused the error (if applicable)",
            "examples": [
              "email",
              "user_id"
            ],
            "type": "string"
          },
          "details": {
            "description": "Additional error context information"
          }
        },
        "required": [
          "type",
          "code",
          "message"
        ],
        "description": "Not found error"
      },
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique object identifier.\nThe format and length of IDs may change over time.",
            "example": "tod_XXXXXXXXXXXXXXXXXXXXXXXXX"
          },
          "userID": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "done"
            ],
            "description": "Status of a todo item.",
            "example": "pending"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "userID",
          "title",
          "status",
          "dueDate"
        ],
        "description": "A todo item that belongs to a user.",
        "example": {
          "id": "tod_XXXXXXXXXXXXXXXXXXXXXXXXX",
          "userID": "usr_XXXXXXXXXXXXXXXXXXXXXXXXX",
          "title": "Write the quarterly report",
          "status": "pending",
          "dueDate": null
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "servers": [
    {
      "description": "Local",
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/todo": {
      "get": {
        "operationId": "getTodo",
        "tags": [
          "Todo"
        ],
        "summary": "List todos",
        "description": "List the current user's todos, optionally filtered by status. Paginated.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Todo"
                      }
                    },
                    "page": {
                      "type": "number",
                      "description": "Page number returned."
                    },
                    "pageSize": {
                      "type": "number",
                      "description": "Number of entities per page."
                    },
                    "total": {
                      "type": "number",
                      "description": "Total number of matching entities across all pages."
                    }
                  },
                  "required": [
                    "data",
                    "page",
                    "pageSize",
                    "total"
                  ],
                  "description": "A page of results."
                },
                "example": {
                  "data": [
                    {
                      "id": "tod_XXXXXXXXXXXXXXXXXXXXXXXXX",
                      "userID": "usr_XXXXXXXXXXXXXXXXXXXXXXXXX",
                      "title": "Write the quarterly report",
                      "status": "pending",
                      "dueDate": null
                    }
                  ],
                  "page": 1,
                  "pageSize": 20,
                  "total": 1
                }
              }
            },
            "description": "A page of todos."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "type": "authentication",
                  "code": "unauthorized",
                  "message": "Authentication required"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "type": "internal",
                  "code": "internal_error",
                  "message": "Internal server error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "in_progress",
                "done"
              ],
              "example": "pending"
            },
            "description": "Status of a todo item."
          }
        ]
      },
      "post": {
        "operationId": "postTodo",
        "tags": [
          "Todo"
        ],
        "summary": "Create todo",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            },
            "description": "The created todo."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "type": "validation",
                  "code": "invalid_parameter",
                  "message": "The request was invalid"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "type": "authentication",
                  "code": "unauthorized",
                  "message": "Authentication required"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "type": "internal",
                  "code": "internal_error",
                  "message": "Internal server error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "dueDate": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}